Revert: API Update for filesystem refactor  [15b491e]

+++ src/SilverStripe/BehatExtension/Context/BasicContext.php
@@ -2,10 +2,13 @@
 
 namespace SilverStripe\BehatExtension\Context;
 
-use Behat\Behat\Context\BehatContext,
+use Behat\Behat\Context\ClosuredContextInterface,
+    Behat\Behat\Context\TranslatedContextInterface,
+    Behat\Behat\Context\BehatContext,
     Behat\Behat\Context\Step,
     Behat\Behat\Event\StepEvent,
-    Behat\Behat\Event\ScenarioEvent;
+    Behat\Behat\Event\ScenarioEvent,
+    Behat\Behat\Exception\PendingException;
 
 use Behat\Mink\Driver\Selenium2Driver;
 
@@ -265,9 +268,8 @@ JS;
      */
 	public function cleanAssetsAfterScenario(ScenarioEvent $event) {
         foreach(\File::get() as $file) {
-            $file->delete();
+            if(file_exists($file->getFullPath())) $file->delete();
         }
-		\Filesystem::removeFolder(ASSETS_PATH, true);
     }
 
     public function takeScreenshot(StepEvent $event) {
